software/python/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "harp-env-sensor"
description = "Driver for Environment Sensor harp device"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Allen Institute for Neural Dynamics"}
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    'pyyaml',
    'pyharp',
    'pyserial',
]

[project.optional-dependencies]
dev = [
    # 'black',
    # 'coverage',
    # 'flake8',
    # 'interrogate',
    # 'isort',
    # 'Sphinx',
    # 'furo',
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "harp_env_sensor.__version__"}